+++ src/main/java/insane96mcp/iguanatweaksexpanded/network/NetworkHandler.java
@@ -21,7 +21,7 @@ public class NetworkHandler {
 		CHANNEL.registerMessage(++index, JumpMidAirMessage.class, JumpMidAirMessage::encode, JumpMidAirMessage::decode, JumpMidAirMessage::handle);
 		CHANNEL.registerMessage(++index, ElectrocutionParticleMessage.class, ElectrocutionParticleMessage::encode, ElectrocutionParticleMessage::decode, ElectrocutionParticleMessage::handle);
 		CHANNEL.registerMessage(++index, SyncITEEnchantingTableStatus.class, SyncITEEnchantingTableStatus::encode, SyncITEEnchantingTableStatus::decode, SyncITEEnchantingTableStatus::handle);
-		CHANNEL.registerMessage(++index, SyncITEEnchantingTableUnlockedEnchantments.class, SyncITEEnchantingTableUnlockedEnchantments::encode, SyncITEEnchantingTableUnlockedEnchantments::decode, SyncITEEnchantingTableUnlockedEnchantments::handle);
+		CHANNEL.registerMessage(++index, SyncITEEnchantingTableLearnedEnchantments.class, SyncITEEnchantingTableLearnedEnchantments::encode, SyncITEEnchantingTableLearnedEnchantments::decode, SyncITEEnchantingTableLearnedEnchantments::handle);
 		CHANNEL.registerMessage(++index, SyncITEEnchantingTableEnchantments.class, SyncITEEnchantingTableEnchantments::encode, SyncITEEnchantingTableEnchantments::decode, SyncITEEnchantingTableEnchantments::handle);
 		CHANNEL.registerMessage(++index, SendMultiBlockFurnaceGhostData.class, SendMultiBlockFurnaceGhostData::encode, SendMultiBlockFurnaceGhostData::decode, SendMultiBlockFurnaceGhostData::handle);
 	}